feat: add slackWebhook as configuration option
This adds slackWebhook as a configuration option with the default value being
the already existing SLACK_WEBHOOK environment variable.

With this, people will be able to provide a slack webhook for this plugin and only
for this plugin, without exposing the webhook through the environment to any
other plugin or program.

Closes issue #16