Allows to use a custom environment variable for Slack webhook
This is particularly useful when using CircleCI and their `context` system.
In our organisation, the context has a default SLACK_WEBHOOK value but we need to override it for some repositories (but still use the context).

Since in CircleCI the context override custom environment variables we need this parameter to use another variable.