feat: allow to override slack channel when using webhook
Currently, there is a logic to throw an error when only `SLACK_CHANNEL`
is specified but not `SLACK_TOKEN`. When using incoming webhook, a
developer can override the slack channel to send a notification to by
specifying the `channel` param for the post request.

This commit changes the logic so that a developer can use specify
`SLACK_WEBHOOK` and `SLACK_CHANNEL` to override the default channel
associated to the webhook.